refactor: bring context down to root of 4.0
