--- src/4.0/validate/index.ts
@@ -1,1 +1,0 @@
-export * from "./context";
